# [Enhancement] Optimize the memory usage of SegmentZonemap in all Null scenarios