--- a/be/src/script/script.cpp
+++ b/be/src/script/script.cpp
@@ -193,16 +193,37 @@ void bind_exec_env(ForeignModule& m) {
     {
         auto& cls = m.klass<GlobalEnv>("GlobalEnv");
         REG_STATIC_METHOD(GlobalEnv, GetInstance);
+
+        // level 0
         REG_METHOD(GlobalEnv, process_mem_tracker);
+
+        // level 1
         REG_METHOD(GlobalEnv, query_pool_mem_tracker);
         REG_METHOD(GlobalEnv, load_mem_tracker);
         REG_METHOD(GlobalEnv, metadata_mem_tracker);
+        REG_METHOD(GlobalEnv, compaction_mem_tracker);
+        REG_METHOD(GlobalEnv, schema_change_mem_tracker);
+        REG_METHOD(GlobalEnv, column_pool_mem_tracker);
+        REG_METHOD(GlobalEnv, page_cache_mem_tracker);
+        REG_METHOD(GlobalEnv, update_mem_tracker);
+        REG_METHOD(GlobalEnv, chunk_allocator_mem_tracker);
+        REG_METHOD(GlobalEnv, clone_mem_tracker);
+        REG_METHOD(GlobalEnv, consistency_mem_tracker);
+
+        // level 2
         REG_METHOD(GlobalEnv, tablet_metadata_mem_tracker);
         REG_METHOD(GlobalEnv, rowset_metadata_mem_tracker);
         REG_METHOD(GlobalEnv, segment_metadata_mem_tracker);
-        REG_METHOD(GlobalEnv, compaction_mem_tracker);
-        REG_METHOD(GlobalEnv, update_mem_tracker);
-        REG_METHOD(GlobalEnv, clone_mem_tracker);
+        REG_METHOD(GlobalEnv, column_metadata_mem_tracker);
+
+        // level 3
+        REG_METHOD(GlobalEnv, tablet_schema_mem_tracker);
+        REG_METHOD(GlobalEnv, column_zonemap_index_mem_tracker);
+        REG_METHOD(GlobalEnv, ordinal_index_mem_tracker);
+        REG_METHOD(GlobalEnv, bitmap_index_mem_tracker);
+        REG_METHOD(GlobalEnv, bloom_filter_index_mem_tracker);
+        REG_METHOD(GlobalEnv, segment_zonemap_mem_tracker);
+        REG_METHOD(GlobalEnv, short_key_index_mem_tracker);
     }
 }