… be/src/storage/rowset/column_reader.cpp