Currently if the column type is varchar(length) and the values is all Null, a string of length will be written to the Segment file, causing the loaded metadata to occupy a large amount of memory. 

This PR optimizes mem usage of reading first, and the memory usage of write and segment file size will be optimization on next pr.

```
CREATE TABLE `t1` (
  `c1` int(11) NULL COMMENT "",
  `c2` varchar(65533) NULL COMMENT "",
  `c3` int(11) NULL COMMENT ""
) ENGINE=OLAP 
DUPLICATE KEY(`c1`, `c2`)
DISTRIBUTED BY HASH(`c1`) BUCKETS 192 
PROPERTIES (
"replication_num" = "1",
"in_memory" = "false",
"enable_persistent_index" = "false",
"replicated_storage" = "true",
"compression" = "LZ4"
);
```

```
mysql> select count(*) from t1;
+-----------+
| count(*)  |
+-----------+
| 143999468 |
+-----------+

```

Before opt: 12M

```
mysql> select count(c2) from t1;                                                                                                                                                                                                                                                  
+-----------+                                                                                                                                                                                                                                                                     
| count(c2) |                                                                                                                                                                                                                                                                     
+-----------+                                                                                                                                                                                                                                                                     
|         0 |                                                                                                                                                                                                                                                                     
+-----------+                                                                                                                                                                                                                                                                     
1 row in set (0.09 sec)                                                                                                                                                                                                                                                           
                                                                                                                                                                                                                                                                                  
mysql> admin execute on 10005 'System.print(GlobalEnv.GetInstance().segment_zonemap_mem_tracker().consumption())';                                                                                                                                                                
+----------+                                                                                                                                                                                                                                                                      
| result   |                                                                                                                                                                                                                                                                      
+----------+                                                                                                                                                                                                                                                                      
| 12647616 |                                                                                                                                                                                                                                                                      
+----------+                                                                                                                                                                                                                                                                      
1 row in set (0.00 sec)   
```

after opt: 73K

```
mysql> select count(c2) from t1;
+-----------+
| count(c2) |
+-----------+
|         0 |
+-----------+
1 row in set (0.08 sec)

mysql> admin execute on 10005 'System.print(GlobalEnv.GetInstance().segment_zonemap_mem_tracker().consumption())';
+--------+
| result |
+--------+
| 73256  |
+--------+
1 row in set (0.01 sec)

```

## What type of PR is this:
- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Checklist:
- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 3.1
  - [x] 3.0
  - [x] 2.5
  - [x] 2.4